requirements.txt
daphne
djangorestframework
django-cors-headers
django
channels